FIX: Huge fix. Code had some issues with the line reporting in errors because of the preprocessing. Took the opportunity to improve the error messages. They now give the line but also the file where it fails.

+++ include/semanticAnalyser/struct.h
@@ -31,7 +31,7 @@ typedef struct SymbolNode {
     SemanticType *param;
 
     SymbolNode *next;
-    int line;
+    size_t line;
 } SymbolNode;
 
 typedef struct ScopeNode {
